requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-genai==1.0.0
chromadb==0.5.0
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.9.2
